Fixed some exception that were not handled if the RawBody was null in the JsonRESTService class.  Also added new constructor options.
